requirements.txt
pycryptodomex>=3.19.0
